+++ tests/test_generate_report_for_configuration.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from pytest_mock import MockFixture
 
-from CLI.support.ablation_help import AblationScenario
+from sparkle.solver.ablation import AblationScenario
 from sparkle.platform import generate_report_for_configuration as sgrch
 from CLI.help import global_variables as gv
 from sparkle.platform.settings_objects import Settings
@@ -103,7 +103,7 @@ def test_get_performance_measure_quality(mocker: MockFixture) -> None:
 
 def test_get_ablation_bool_true(mocker: MockFixture) -> None:
     """Test get_ablation_bool returns correct string if get_ablation_bool is True."""
-    mock_check = mocker.patch("CLI.support.ablation_help.AblationScenario."
+    mock_check = mocker.patch("sparkle.solver.ablation.AblationScenario."
                               "check_for_ablation",
                               return_value=True)
     ablation_bool = sgrch.get_ablation_bool(ablation_scenario)
@@ -113,7 +113,7 @@ def test_get_ablation_bool_true(mocker: MockFixture) -> None:
 
 def test_get_ablation_bool_false(mocker: MockFixture) -> None:
     """Test get_ablation_bool returns correct string if get_ablation_bool is False."""
-    mock_check = mocker.patch("CLI.support.ablation_help.AblationScenario."
+    mock_check = mocker.patch("sparkle.solver.ablation.AblationScenario."
                               "check_for_ablation",
                               return_value=False)
     ablation_bool = sgrch.get_ablation_bool(ablation_scenario)
@@ -291,7 +291,7 @@ def test_get_ablation_table(mocker: MockFixture) -> None:
          ["0", "-source-", "N/A", "N/A", "76.53275"],
          ["1", "sel_var_div", "3", "6", "68.41392"],
          ["2", "-target-", "N/A", "N/A", "92.06944"]])
-    mock_table = mocker.patch("CLI.support.ablation_help.AblationScenario."
+    mock_table = mocker.patch("sparkle.solver.ablation.AblationScenario."
                               "read_ablation_table",
                               return_value=sah_ablation_table)
